+++ src/App/VRChatContentPublisher.App/Extensions/ServicesExtenstion.cs
@@ -1,7 +1,9 @@
 ﻿using Microsoft.Extensions.DependencyInjection;
 using VRChatContentPublisher.App.Services;
+using VRChatContentPublisher.App.Services.AppLifetime;
 using VRChatContentPublisher.App.Services.Dialog;
-using VRChatContentPublisher.App.Services.NotificationSender;
+using VRChatContentPublisher.App.Services.Notification;
+using VRChatContentPublisher.App.Services.Notification.Sender;
 using VRChatContentPublisher.App.Services.Update;
 using VRChatContentPublisher.App.ViewModels;
 using VRChatContentPublisher.App.ViewModels.Data;
@@ -30,10 +32,13 @@ public static class ServicesExtenstion
 
         services.AddSingleton<AppLifetimeService>();
         services.AddSingleton<AppWindowService>();
-        services.AddSingleton<AppNotificationService>();
         services.AddSingleton<IActivateWindowService>(s => s.GetRequiredService<AppWindowService>());
 
         services.AddSingleton<InAppNotificationService>();
+
+        // Notification
+        services.AddHostedService<AppNotificationHostedService>();
+        services.AddSingleton<DesktopNotificationService>();
 
         // In App Notification
         services.AddTransient<PublicIpChangedInAppNotificationViewModelFactory>();